[testsuite] add x86 effective target
I got tired of repeating the conditional that recognizes ia32 or
x86_64, and introduced 'x86' as a shorthand for that, adjusting all
occurrences in target-supports.exp, to set an example.  I found some
patterns that recognized i?86* and x86_64*, but I took those as likely
cut&pastos instead of trying to preserve those weirdnesses.


for  gcc/ChangeLog

	* doc/sourcebuild.texi: Add x86 effective target.

for  gcc/testsuite/ChangeLog

	* lib/target-supports.exp (check_effective_target_x86): New.
	Replace all uses of i?86-*-* and x86_64-*-* in this file.