testsuite: arm: Avoid incremental link warnings in pr61123-enum-size
This test uses incremental linking, but that can generate warnings if
the LTO step contains a mix of LTO and non-LTO object files (this can
happen when there's a testglue file that is normally included during
linking).

We don't care about the testglue, though, so just tell the LTO
optimizer to generate nolto-rel output, which is what it is falling
back to anyway.

gcc/testsuite:
	* gcc.target/arm/lto/pr61123-enum-size_0.c: (dg-lto-options) Move
	linker related options to ...
	(dg-extra-ld-options): ... here.  Add -flinker-output=nolto-rel.