testsuite: Fix up gcc.target/i386/pr118940.c test [PR118940]
The testcase uses -m32 in dg-options, something we try hard not to do,
if something should be tested only for -m32, it is { target ia32 } test,
if it can be tested for -m64/-mx32 too, just some extra options are
needed for ia32, it should have dg-additional-options with ia32 target.

Also, the test wasn't reduced, so I've reduced it using cvise and manual
tweaks and verified the test still FAILs before r15-7700 and succeeds
with current trunk.

2025-02-28  Jakub Jelinek  <jakub@redhat.com>

	PR target/118940
	* gcc.target/i386/pr118940.c: Drop -w, -g and -m32 from dg-options, move
	-march=i386 -mregparm=3 to dg-additional-options for ia32 and -fno-pie
	to dg-additional-options for pie.  Reduce the test.